fix: handle no raw mode
